+++ src/anglrFileFormatter.ts
@@ -5,7 +5,7 @@ import {Project, SourceFile, FormatCodeSettings} from 'ts-morph';
 
 import {FormatterType} from './formatters/formatters.interface';
 import {NewLineType} from './misc';
-import {ImportFormatter, DecoratorFormatter} from './formatters';
+import {ImportFormatter, DecoratorFormatter, CallExpressionFormatter, ConstructorFormatter} from './formatters';
 
 /**
  * Class that represents instance of formatter for one file
@@ -57,6 +57,8 @@ export class AnglrFileFormatter
         //registers formatters
         this._formatters.push(ImportFormatter);
         this._formatters.push(DecoratorFormatter);
+        this._formatters.push(ConstructorFormatter);
+        this._formatters.push(CallExpressionFormatter);
     }
 
     //######################### public methods #########################