+++ src/formatters/index.ts
@@ -1,2 +1,4 @@
 export * from './decorator.formatter';
 export * from './import.formatter';
+export * from './callExpression.formatter';
+export * from './constructor.formatter';