+++ changelog.md
@@ -9,3 +9,4 @@
 - *formatter*
     - supports reformating of *import* statements
     - supports reformating of *decorator* declarations
+    - supports reformating of *constructor* parameters declaration